chore: reorganize checkout session tests
